Rename src/benchmarks -> src/chain
