+++ test/benchmarks/02-erc20-transfers.test.ts
@@ -2,8 +2,8 @@ import { expect } from 'chai'
 import { Wallet } from 'ethers'
 import { Contract } from 'ethers/contract'
 import { parseEther } from 'ethers/utils'
-import { SimpleProvider } from '../../src/benchmarks/SimpleProvider'
-import { createSimpleChain, SimpleChain } from '../../src/benchmarks/SimpleChain'
+import { SimpleProvider } from '../../src/chain/SimpleProvider'
+import { createSimpleChain, SimpleChain } from '../../src/chain/SimpleChain'
 import { getERC20TransferTransaction } from './utils/transactions'
 import { deployERC20 } from './utils/deploy'
