+++ test/benchmarks/03-erc20-deploys.test.ts
@@ -1,8 +1,8 @@
 import { expect } from 'chai'
 import { parseEther } from 'ethers/utils'
 import { Contract } from 'ethers/contract'
-import { createSimpleChain, SimpleChain } from '../../src/benchmarks/SimpleChain'
-import { SimpleProvider } from '../../src/benchmarks/SimpleProvider'
+import { createSimpleChain, SimpleChain } from '../../src/chain/SimpleChain'
+import { SimpleProvider } from '../../src/chain/SimpleProvider'
 import { getERC20DeploymentTransaction } from './utils/transactions'
 import { deployERC20 } from './utils/deploy'
