+++ test/benchmarks/06-many-storage.test.ts
@@ -1,7 +1,7 @@
 import { Wallet } from 'ethers'
 import { Contract } from 'ethers/contract'
-import { SimpleProvider } from '../../src/benchmarks/SimpleProvider'
-import { createSimpleChain, SimpleChain } from '../../src/benchmarks/SimpleChain'
+import { SimpleProvider } from '../../src/chain/SimpleProvider'
+import { createSimpleChain, SimpleChain } from '../../src/chain/SimpleChain'
 import { deployERC20 } from './utils/deploy'
 import { getERC20ApproveTransaction, getERC20TransferTransaction } from './utils/transactions'
 import { parseEther } from 'ethers/utils'
